[WIP]: Moves search document building code to separate file
